Fixes #4168 Remove User TagCategory initialized during system startup
