+++ catalog-rest-service/src/test/java/org/openmetadata/catalog/resources/glossary/GlossaryResourceTest.java
@@ -31,12 +31,17 @@ import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.openmetadata.catalog.Entity;
 import org.openmetadata.catalog.api.data.CreateGlossary;
+import org.openmetadata.catalog.api.data.CreateGlossaryTerm;
 import org.openmetadata.catalog.entity.data.Glossary;
+import org.openmetadata.catalog.entity.data.GlossaryTerm;
 import org.openmetadata.catalog.jdbi3.GlossaryRepository.GlossaryEntityInterface;
+import org.openmetadata.catalog.jdbi3.GlossaryTermRepository.GlossaryTermEntityInterface;
 import org.openmetadata.catalog.resources.EntityResourceTest;
 import org.openmetadata.catalog.type.ChangeDescription;
 import org.openmetadata.catalog.type.EntityReference;
 import org.openmetadata.catalog.type.FieldChange;
+import org.openmetadata.catalog.type.TagLabel;
+import org.openmetadata.catalog.type.TagLabel.Source;
 import org.openmetadata.catalog.util.EntityInterface;
 import org.openmetadata.catalog.util.JsonUtils;
 import org.openmetadata.catalog.util.TestUtils;
@@ -51,6 +56,43 @@ public class GlossaryResourceTest extends EntityResourceTest<Glossary, CreateGlo
   @BeforeAll
   public void setup(TestInfo test) throws IOException, URISyntaxException {
     super.setup(test);
+  }
+
+  public void setupGlossaries() throws HttpResponseException {
+    GlossaryResourceTest glossaryResourceTest = new GlossaryResourceTest();
+    CreateGlossary createGlossary = glossaryResourceTest.createRequest("g1", "", "", null);
+    GLOSSARY1 = glossaryResourceTest.createEntity(createGlossary, ADMIN_AUTH_HEADERS);
+    GLOSSARY1_REF = new GlossaryEntityInterface(GLOSSARY1).getEntityReference();
+
+    createGlossary = glossaryResourceTest.createRequest("g2", "", "", null);
+    GLOSSARY2 = glossaryResourceTest.createEntity(createGlossary, ADMIN_AUTH_HEADERS);
+    GLOSSARY2_REF = new GlossaryEntityInterface(GLOSSARY2).getEntityReference();
+
+    GlossaryTermResourceTest glossaryTermResourceTest = new GlossaryTermResourceTest();
+    CreateGlossaryTerm createGlossaryTerm =
+        glossaryTermResourceTest
+            .createRequest("g1t1", null, "", null)
+            .withRelatedTerms(null)
+            .withGlossary(GLOSSARY1_REF);
+    GLOSSARY1_TERM1 = glossaryTermResourceTest.createEntity(createGlossaryTerm, ADMIN_AUTH_HEADERS);
+    GLOSSARY1_TERM1_REF = new GlossaryTermEntityInterface(GLOSSARY1_TERM1).getEntityReference();
+    GLOSSARY1_TERM1_LABEL = getTagLabel(GLOSSARY1_TERM1);
+
+    createGlossaryTerm =
+        glossaryTermResourceTest
+            .createRequest("g2t1", null, "", null)
+            .withRelatedTerms(null)
+            .withGlossary(GLOSSARY2_REF);
+    GLOSSARY2_TERM1 = glossaryTermResourceTest.createEntity(createGlossaryTerm, ADMIN_AUTH_HEADERS);
+    GLOSSARY2_TERM1_REF = new GlossaryTermEntityInterface(GLOSSARY2_TERM1).getEntityReference();
+    GLOSSARY2_TERM1_LABEL = getTagLabel(GLOSSARY2_TERM1);
+  }
+
+  private TagLabel getTagLabel(GlossaryTerm term) {
+    return new TagLabel()
+        .withTagFQN(term.getFullyQualifiedName())
+        .withDescription(term.getDescription())
+        .withSource(Source.GLOSSARY);
   }
 
   @Test
