+++ catalog-rest-service/src/test/java/org/openmetadata/catalog/resources/services/DashboardServiceResourceTest.java
@@ -23,16 +23,22 @@ import static org.openmetadata.catalog.util.TestUtils.getPrincipal;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.http.client.HttpResponseException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
 import org.openmetadata.catalog.Entity;
+import org.openmetadata.catalog.api.data.CreateChart;
 import org.openmetadata.catalog.api.services.CreateDashboardService;
+import org.openmetadata.catalog.api.services.CreateDashboardService.DashboardServiceType;
+import org.openmetadata.catalog.entity.data.Chart;
 import org.openmetadata.catalog.entity.services.DashboardService;
+import org.openmetadata.catalog.jdbi3.ChartRepository.ChartEntityInterface;
 import org.openmetadata.catalog.jdbi3.DashboardServiceRepository.DashboardServiceEntityInterface;
 import org.openmetadata.catalog.resources.EntityResourceTest;
+import org.openmetadata.catalog.resources.charts.ChartResourceTest;
 import org.openmetadata.catalog.resources.services.dashboard.DashboardServiceResource.DashboardServiceList;
 import org.openmetadata.catalog.services.connections.dashboard.SupersetConnection;
 import org.openmetadata.catalog.type.ChangeDescription;
@@ -221,6 +227,27 @@ public class DashboardServiceResourceTest extends EntityResourceTest<DashboardSe
         assertEquals(expectedSupersetConnection.getPassword(), actualSupersetConnection.getPassword());
         assertEquals(expectedSupersetConnection.getProvider(), actualSupersetConnection.getProvider());
       }
+    }
+  }
+
+  public void setupDashboardServices(TestInfo test) throws HttpResponseException {
+    DashboardServiceResourceTest dashboardResourceTest = new DashboardServiceResourceTest();
+    CreateDashboardService createDashboardService =
+        dashboardResourceTest.createRequest("superset", "", "", null).withServiceType(DashboardServiceType.Superset);
+
+    DashboardService dashboardService =
+        new DashboardServiceResourceTest().createEntity(createDashboardService, ADMIN_AUTH_HEADERS);
+    SUPERSET_REFERENCE = new DashboardServiceEntityInterface(dashboardService).getEntityReference();
+
+    createDashboardService.withName("looker").withServiceType(DashboardServiceType.Looker);
+    dashboardService = new DashboardServiceResourceTest().createEntity(createDashboardService, ADMIN_AUTH_HEADERS);
+    LOOKER_REFERENCE = new DashboardServiceEntityInterface(dashboardService).getEntityReference();
+    CHART_REFERENCES = new ArrayList<>();
+    ChartResourceTest chartResourceTest = new ChartResourceTest();
+    for (int i = 0; i < 3; i++) {
+      CreateChart createChart = chartResourceTest.createRequest(test, i).withService(SUPERSET_REFERENCE);
+      Chart chart = chartResourceTest.createEntity(createChart, ADMIN_AUTH_HEADERS);
+      CHART_REFERENCES.add(new ChartEntityInterface(chart).getEntityReference());
     }
   }
 }
